🗂️ Reorganize geometry module for consistency and clarity
Implements consistent file organization across all geometry components
following clear naming conventions and separation of concerns.

## Boundary Conditions: Consistent Separation
- Created `boundary_conditions_1d.py` (extracted from domain_1d.py)
- Now all dimensions follow same pattern:
  - boundary_conditions_1d.py (6.5K)
  - boundary_conditions_2d.py (25K)
  - boundary_conditions_3d.py (24K)
- Updated domain_1d.py to import BCs instead of defining them
- Maintains backward compatibility through __init__.py exports

## AMR Files: Clear Prefix and Dimension Suffix
Renamed all AMR files with `amr_` prefix and dimension suffix:
- one_dimensional_amr.py → amr_1d.py
- amr_mesh.py → amr_quadtree_2d.py (structured quadtree)
- triangular_amr.py → amr_triangular_2d.py (unstructured triangular)
- tetrahedral_amr.py → amr_tetrahedral_3d.py

Benefits:
- Visual file grouping (all amr_* files together)
- Clear dimension identification (1d/2d/3d suffix)
- Descriptive names showing refinement strategy

## Import Updates
- Updated __init__.py to import from renamed files
- Fixed cross-references in amr_1d.py and amr_triangular_2d.py
- All imports tested and working correctly

## Final Structure
```
boundary_conditions_1d.py, _2d.py, _3d.py  (BCs separated by dimension)
domain_1d.py, _2d.py, _3d.py                (Domains separated by dimension)
amr_1d.py                                   (1D interval refinement)
amr_quadtree_2d.py                          (2D structured refinement)
amr_triangular_2d.py                        (2D unstructured refinement)
amr_tetrahedral_3d.py                       (3D tetrahedral refinement)
```

This organization provides clear separation, consistent naming, and
better file navigation while maintaining full backward compatibility.
