+++ mfg_pde/geometry/amr_1d.py
@@ -15,7 +15,7 @@ from typing import TYPE_CHECKING, Any
 
 import numpy as np
 
-from .amr_mesh import AMRRefinementCriteria, BaseErrorEstimator
+from .amr_quadtree_2d import AMRRefinementCriteria, BaseErrorEstimator
 from .base_geometry import MeshData
 
 logger = logging.getLogger(__name__)
